yolo5/requirements.txt
flask
pyyaml
loguru
#Add boto3 - AWS SDK for python
boto3